VN upgrade, some more refactoring of PetService, Beehive WIP

+++ src/Controller/Item/BugController.php
@@ -4,12 +4,16 @@ namespace App\Controller\Item;
 use App\Entity\Inventory;
 use App\Entity\PetActivityLog;
 use App\Enum\LocationEnum;
+use App\Enum\SerializationGroupEnum;
+use App\Enum\StoryEnum;
 use App\Enum\UserStatEnum;
 use App\Repository\InventoryRepository;
 use App\Repository\ItemRepository;
+use App\Repository\UserQuestRepository;
 use App\Repository\UserStatsRepository;
 use App\Service\InventoryService;
 use App\Service\ResponseService;
+use App\Service\StoryService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
@@ -27,12 +31,17 @@ class BugController extends PoppySeedPetsItemController
      */
     public function squishBug(
         Inventory $inventory, ResponseService $responseService, UserStatsRepository $userStatsRepository,
-        EntityManagerInterface $em
+        EntityManagerInterface $em, UserQuestRepository $userQuestRepository
     )
     {
         $user = $this->getUser();
 
         $this->validateInventory($inventory, 'bug/#/squish');
+
+        $promised = $userQuestRepository->findOrCreate($user, 'Promised to Not Squish Bugs', 0);
+
+        if($promised->getValue())
+            return $responseService->itemActionSuccess('You\'ve promised not to squish any more bugs...');
 
         $em->remove($inventory);
 
@@ -141,5 +150,24 @@ class BugController extends PoppySeedPetsItemController
         $responseService->addActivityLog((new PetActivityLog())->setEntry($message));
 
         return $responseService->itemActionSuccess(null, [ 'reloadInventory' => true, 'itemDeleted' => true ]);
+    }
+
+    /**
+     * @Route("/{inventory}/talkToQueen", methods={"POST"})
+     * @IsGranted("IS_AUTHENTICATED_FULLY")
+     * @throws \Exception
+     */
+    public function talkToQueen(
+        Inventory $inventory, StoryService $storyService, Request $request, EntityManagerInterface $em,
+        ResponseService $responseService
+    )
+    {
+        $user = $this->getUser();
+
+        $this->validateInventory($inventory, 'bug/#/squish');
+
+        $response = $storyService->doStory($user, StoryEnum::STOLEN_PLANS, $request->request);
+
+        return $responseService->success($response, SerializationGroupEnum::STORY);
     }
 }
